requirements.txt
flask
opencv-python
pillow
rembg
qrcode
